// admin-ui/src/generated-api-types/models/Media.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Optional object for event-related media (image or video).
 */
export type Media = {
    type: 'image' | 'video';
    /**
     * URL, CID, or Data URI.
     */
    value: string;
};